finalize theme fixes and restore mood static export route
