+++ components/MoodPieChart.tsx
@@ -45,15 +45,18 @@ export default function MoodPieChart({ data }: MoodPieChartProps) {
             nameKey="name"
             cx="50%"
             cy="50%"
-            outerRadius={100}
-            label={({ name, percent }) => `${name} ${(percent * 100).toFixed(0)}%`}
-            labelLine={true}
+            innerRadius={60}
+            outerRadius={102}
+            label={false}
+            labelLine={false}
+            paddingAngle={2}
           >
             {chartData.map((entry, index) => (
               <Cell key={`cell-${index}`} fill={entry.color} />
             ))}
           </Pie>
           <Tooltip
+            formatter={(value: number, name: string) => [`${value} entries`, name]}
             contentStyle={{
               backgroundColor: 'hsl(var(--popover))',
               color: 'hsl(var(--popover-foreground))',
@@ -62,7 +65,12 @@ export default function MoodPieChart({ data }: MoodPieChartProps) {
               boxShadow: '0 4px 6px -1px rgb(0 0 0 / 0.1)'
             }}
           />
-          <Legend verticalAlign="bottom" height={36} iconType="circle" />
+          <Legend
+            verticalAlign="bottom"
+            iconType="circle"
+            wrapperStyle={{ paddingTop: 14, fontSize: '13px' }}
+            formatter={(value) => <span style={{ color: 'hsl(var(--foreground))' }}>{value}</span>}
+          />
         </PieChart>
       </ResponsiveContainer>
     </div>
