+++ next.config.js
@@ -6,12 +6,7 @@ if (isGithubActions) {
   repoName = `/${repo.split('/')[1]}`;
 }
 
-const withPWA = require('next-pwa')({
-  dest: 'public',
-  register: true,
-  skipWaiting: true,
-  disable: process.env.NODE_ENV === 'development',
-})
+const withPWA = require('next-pwa');
 
 const nextConfig = {
   reactStrictMode: true,
@@ -28,6 +23,12 @@ const nextConfig = {
     'three',
   ],
   outputFileTracingRoot: __dirname,
+  pwa: {
+    dest: 'public',
+    register: true,
+    skipWaiting: true,
+    disable: process.env.NODE_ENV === 'development',
+  },
 };
 
 module.exports = withPWA(nextConfig)
